# CallMonitor/public/openapi.yaml
openapi: 3.0.3
info:
  title: Wordis Bond API
  description: |
    Wordis Bond - Conversation System of Record API
    
    This API provides endpoints for managing voice calls, recordings, transcriptions,
    campaigns, analytics, and compliance features for enterprise conversation management.
    
    ## Authentication
    All endpoints require authentication via NextAuth.js session cookies.
    Include `credentials: 'include'` in all fetch requests.
    
    ## RBAC Roles
    - **owner**: Full access including billing and team management
    - **admin**: Manage team, calls, settings (no billing)
    - **operator**: Make calls, view recordings, run campaigns
    - **analyst**: View recordings, transcripts, analytics
    - **viewer**: View-only access to dashboards
    
    ## Rate Limiting
    API endpoints are rate limited. Standard limits:
    - Read operations: 100 requests/minute
    - Write operations: 30 requests/minute
    - Sensitive operations: 5 requests/minute
  version: 1.0.0
  contact:
    name: Wordis Bond Support
    email: support@wordisbond.com
  license:
    name: Proprietary
    url: https://wordisbond.com/terms
servers:
  - url: /api
    description: API Base URL
...
paths:
...
  /retention/legal-holds:
    get:
      tags:
        - Compliance
      summary: List legal holds
      operationId: listLegalHolds
      responses:
        '200':
          description: Legal holds
          content:
            application/json:
              schema:
                type: object
                properties:
                  legal_holds:
                    type: array
                    items:
                      $ref: '#/components/schemas/LegalHold'
    post:
      tags:
        - Compliance
      summary: Create legal hold
      operationId: createLegalHold
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - hold_name
              properties:
                hold_name:
                  type: string
                matter_reference:
                  type: string
                applies_to_all:
                  type: boolean
      responses:
        '201':
          description: Legal hold created
    delete:
      tags:
        - Compliance
      summary: Release legal hold
      operationId: releaseLegalHold
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - hold_id
                - release_reason
              properties:
                hold_id:
                  type: string
                  format: uuid
                release_reason:
                  type: string
      responses:
        '200':
          description: Legal hold released
...
  /reports:
    get:
      tags:
        - Analytics
      summary: List generated reports
      operationId: listReports
      responses:
        '200':
          description: Reports list
    post:
      tags:
        - Analytics
      summary: Generate new report
      operationId: generateReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - template_id
              properties:
                template_id:
                  type: string
                  format: uuid
                parameters:
                  type: object
      responses:
        '201':
          description: Report generation started
...
components:
  schemas:
...
    LegalHold:
      type: object
      properties:
        id:
          type: string
          format: uuid
        hold_name:
          type: string
        matter_reference:
          type: string
        status:
          type: string
          enum: [active, released]
        applies_to_all:
          type: boolean
        created_at:
          type: string
          format: date-time
        created_by:
          type: string
          format: uuid
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie
security:
  - cookieAuth: []